Refactor feature toggles and exp config types
Changed expIncrement and expVariation types from long to int for consistency. Refactored feature toggling to use disabledFeatures instead of enabledFeatures in guildSettings and userSettings, updating related logic throughout the codebase. Updated globalLevelling to use config values for exp and level parameters. Added command disabling check in eventListeners.

+++ src/main/java/com/jami/botAdmin/commandsAdmin.java
@@ -93,7 +93,7 @@ public class commandsAdmin {
           return;
         }
         try {
-          App.CONFIG.setExpIncrement(Long.valueOf(args.get(1)));
+          App.CONFIG.setExpIncrement(Integer.valueOf(args.get(1)));
           event.getMessage().reply("Exp Increment set to " + args.get(1)).queue();
         } catch (Exception e) {
           event.getMessage().reply("**Command Usage:** a!set-exp-increment <exp>").queue();
@@ -106,7 +106,7 @@ public class commandsAdmin {
           return;
         }
         try {
-          App.CONFIG.setExpVariation(Long.valueOf(args.get(1)));
+          App.CONFIG.setExpVariation(Integer.valueOf(args.get(1)));
           event.getMessage().reply("Exp Variation set to " + args.get(1)).queue();
         } catch (Exception e) {
           event.getMessage().reply("**Command Usage:** a!set-exp-variation <variation>").queue();
